+++ integration/graphql_types_test.go
@@ -128,8 +128,8 @@ type Party struct {
 	Proposals          []Proposal
 	Votes              []Vote
 	LiquidityProvision []LiquidityProvision
+	Positions          []Position
 	// TODO:
-	// Positions []Position
 	// Margins []MarginLevels
 	Withdrawals []Withdrawal
 	Deposits    []Deposit
@@ -370,4 +370,14 @@ type OracleSpec struct {
 	Filters   []Filter
 	Status    OracleSpecStatus
 	Data      []OracleData
+}
+
+type Position struct {
+	Market            Market
+	Party             Party
+	OpenVolume        string
+	RealisedPNL       string
+	UnrealisedPNL     string
+	AverageEntryPrice string
+	UpdatedAt         TimeString
 }